♻️ refactor: extract Home row/card into standalone subviews
Split HomeScenarioRow / HomeScenarioMetaLine / HomePausedCard out of
HomeView into their own View types, and break scenarioList into
pausedSection / scenariosSection helpers. Pure structural extraction (no
behavior change) to bring HomeView back under SwiftLint file_length /
type_body_length / function_body_length limits after the P2 additions.
